# Create CucumberExpression instances once