--- a/src/runtime/test_case_runner.js
+++ b/src/runtime/test_case_runner.js
@@ -105,7 +105,6 @@ export default class TestCaseRunner {
     return this.supportCodeLibrary.stepDefinitions.filter(stepDefinition =>
       stepDefinition.matchesStepName({
         stepName: step.text,
-        parameterTypeRegistry: this.supportCodeLibrary.parameterTypeRegistry,
       })
     )
   }
@@ -114,7 +113,6 @@ export default class TestCaseRunner {
     return StepRunner.run({
       defaultTimeout: this.supportCodeLibrary.defaultTimeout,
       hookParameter,
-      parameterTypeRegistry: this.supportCodeLibrary.parameterTypeRegistry,
       step,
       stepDefinition,
       world: this.world,