--- a/src/support_code_library_builder/build_helpers.js
+++ b/src/support_code_library_builder/build_helpers.js
@@ -1,10 +1,14 @@
 import { deprecate } from 'util'
 import _ from 'lodash'
 import { formatLocation } from '../formatter/helpers'
-import { ParameterType } from 'cucumber-expressions'
+import {
+  ParameterType,
+  CucumberExpression,
+  RegularExpression,
+} from 'cucumber-expressions'
 import path from 'path'
 import StackTrace from 'stacktrace-js'
-import UnboundStepDefinition from '../models/unbound_step_definition'
+import StepDefinition from '../models/step_definition'
 import TestCaseHookDefinition from '../models/test_case_hook_definition'
 import TestRunHookDefinition from '../models/test_run_hook_definition'
 import validateArguments from './validate_arguments'
@@ -51,7 +55,7 @@ export function buildTestRunHookDefinition({ options, code, cwd }) {
   })
 }
 
-export function buildStepDefinition({ pattern, options, code, cwd }) {
+export function buildStepDefinitionStruct({ pattern, options, code, cwd }) {
   if (typeof options === 'function') {
     code = options
     options = {}
@@ -62,13 +66,22 @@ export function buildStepDefinition({ pattern, options, code, cwd }) {
     fnName: 'defineStep',
     location: formatLocation({ line, uri }),
   })
-  return new UnboundStepDefinition({
+  return {
     code,
     line,
     options,
     pattern,
     uri,
-  })
+  }
+}
+
+export function buildStepDefinitionFromStruct(struct, parameterTypeRegistry) {
+  const { code, line, options, uri, pattern } = struct
+  const Expression =
+    typeof pattern === 'string' ? CucumberExpression : RegularExpression
+
+  const expression = new Expression(pattern, parameterTypeRegistry)
+  return new StepDefinition({ code, line, options, uri, pattern, expression })
 }
 
 const projectPath = path.join(__dirname, '..', '..')