I noticed that `--fail-fast` gets slow on big projects.

It appears that this is a result of instantiating `CucumberExpression` instances too many times: once per test of whether the text of a step matches a given step definition (instead of once per _creation_ of that step definition).

This refactoring creates the expression (either a `CucumberExpression` or a `RegularExpression`) once per step definition, during the `finalize()` phase of building the support code library.

The result is much faster `--fail-fast` on large projects.

For example, on my current project, when the first step (of 1103 steps) fails:

Before:
```
> time ./node_modules/.bin/cucumber-js --fail-fast
./node_modules/.bin/cucumber-js --fail-fast  10.24s user 0.37s system 102% cpu 10.342 total
```

After:
```
> time ./node_modules/.bin/cucumber-js --fail-fast
./node_modules/.bin/cucumber-js --fail-fast  2.36s user 0.28s system 109% cpu 2.429 total
```

I also noticed that the various subclasses of `StepDefinition` actually had different properties and methods, so I extracted a `Definition` superclass to isolate the actual commonality.