@aslakhellesoy 2018-11-03
Blimey - great work @joshski!

We’ll merge this as soon as the other implementations are uptodate:

- [x] JavaScript
- [ ] Java
- [ ] Ruby
- [ ] Go