@aslakhellesoy 2018-11-03
Oh hang on I was a bit quick there, this is not in the library itself, but in Cucumber.js.

We can ignore the langauage porting then.