@joshski 2018-11-03
CPU bound test runs appear to be generally faster as a result of this change too... e.g. on the same project, running all the scenarios with in-memory dependencies:

Before:
```
11.31s user 0.40s system 103% cpu 11.299 total
```

After:
```
4.41s user 0.32s system 110% cpu 4.278 total
```