@charlierudolph 2018-11-05
Very cool. I’ll take a closer look at this soon, but sounds like a great idea