@joshski 2018-11-21
I've been using this change for a couple of weeks and it noticeably improves performance, not just with `--fail-fast`, but in general. I'm finding it hard to use cucumber-js without it now!

Is there an easy way to point the `package.json` in my project at this branch (considering the JavaScript is transpiled before releasing to NPM)?