@joshski 2018-11-27
Renames done. I agree those names are better, thanks